Add AbstractMachineBlockEntity
