+++ src/main/java/com/teamofelectrorealism/electrorealism/block/ModBlockEntityTypes.java
@@ -2,11 +2,11 @@ package com.teamofelectrorealism.electrorealism.block;
 
 import com.teamofelectrorealism.electrorealism.ElectroRealism;
 import com.teamofelectrorealism.electrorealism.block.connector.small.SmallConnectorBlockEntity;
-import com.teamofelectrorealism.electrorealism.block.user.crusher.ElectricCrusherBlockEntity;
-import com.teamofelectrorealism.electrorealism.block.user.arc_furnace.ArcFurnaceBlockEntity;
-import com.teamofelectrorealism.electrorealism.block.generator.solarpanel.SolarPanelBlockEntity;
-import com.teamofelectrorealism.electrorealism.block.generator.test.VoltageSourceBlockEntity;
-import com.teamofelectrorealism.electrorealism.block.generator.waterwheel.WaterWheelBlockEntity;
+import com.teamofelectrorealism.electrorealism.block.machine.user.crusher.ElectricCrusherBlockEntity;
+import com.teamofelectrorealism.electrorealism.block.machine.user.arc_furnace.ArcFurnaceBlockEntity;
+import com.teamofelectrorealism.electrorealism.block.machine.generator.solarpanel.SolarPanelBlockEntity;
+import com.teamofelectrorealism.electrorealism.block.machine.generator.test.VoltageSourceBlockEntity;
+import com.teamofelectrorealism.electrorealism.block.machine.generator.waterwheel.WaterWheelBlockEntity;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.neoforged.bus.api.IEventBus;
