+++ src/main/java/com/teamofelectrorealism/electrorealism/block/machine/AbstractMachineBlock.java
@@ -1,4 +1,5 @@
-package com.teamofelectrorealism.electrorealism.block.generator;
+package com.teamofelectrorealism.electrorealism.block.machine;
+
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.BaseEntityBlock;
@@ -9,8 +10,8 @@ import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.minecraft.world.level.block.state.BlockState;
 import org.jetbrains.annotations.Nullable;
 
-public abstract class AbstractGeneratorBlock extends BaseEntityBlock {
-    protected AbstractGeneratorBlock(Properties properties) {
+public abstract class AbstractMachineBlock extends BaseEntityBlock {
+    protected AbstractMachineBlock(Properties properties) {
         super(properties);
     }
 
@@ -22,6 +23,5 @@ public abstract class AbstractGeneratorBlock extends BaseEntityBlock {
     @Override
     public abstract @Nullable <T extends BlockEntity> BlockEntityTicker<T> getTicker(Level level, BlockState state, BlockEntityType<T> blockEntityType);
 
-
-    protected abstract void tick(Level level1, BlockPos pos, BlockState state1, AbstractGeneratorBlockEntity blockEntity);
+    protected abstract void tick(Level level1, BlockPos pos, BlockState state1, AbstractMachineBlockEntity blockEntity);
 }
