+++ src/main/java/com/teamofelectrorealism/electrorealism/block/machine/AbstractMachineBlockEntity.java
@@ -1,0 +1,15 @@
+package com.teamofelectrorealism.electrorealism.block.machine;
+
+import net.minecraft.core.BlockPos;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.entity.BlockEntity;
+import net.minecraft.world.level.block.entity.BlockEntityType;
+import net.minecraft.world.level.block.state.BlockState;
+
+public abstract class AbstractMachineBlockEntity extends BlockEntity {
+    public AbstractMachineBlockEntity(BlockEntityType<?> type, BlockPos pos, BlockState blockState) {
+        super(type, pos, blockState);
+    }
+
+    public abstract void tick(Level level, BlockPos pos, BlockState state);
+}
