+++ src/main/java/com/teamofelectrorealism/electrorealism/block/machine/generator/AbstractGeneratorBlock.java
@@ -1,0 +1,9 @@
+package com.teamofelectrorealism.electrorealism.block.machine.generator;
+
+import com.teamofelectrorealism.electrorealism.block.machine.AbstractMachineBlock;
+
+public abstract class AbstractGeneratorBlock extends AbstractMachineBlock {
+    protected AbstractGeneratorBlock(Properties properties) {
+        super(properties);
+    }
+}
