+++ src/main/java/com/teamofelectrorealism/electrorealism/block/machine/user/AbstractPowerUserBlock.java
@@ -1,0 +1,9 @@
+package com.teamofelectrorealism.electrorealism.block.machine.user;
+
+import com.teamofelectrorealism.electrorealism.block.machine.AbstractMachineBlock;
+
+public abstract class AbstractPowerUserBlock extends AbstractMachineBlock {
+    protected AbstractPowerUserBlock(Properties properties) {
+        super(properties);
+    }
+}
