+++ src/main/java/com/teamofelectrorealism/electrorealism/block/machine/user/crusher/ElectricCrusherBlock.java
@@ -1,9 +1,9 @@
-package com.teamofelectrorealism.electrorealism.block.user.arc_furnace;
+package com.teamofelectrorealism.electrorealism.block.machine.user.crusher;
 
 import com.mojang.serialization.MapCodec;
 import com.teamofelectrorealism.electrorealism.block.ModBlockEntityTypes;
-import com.teamofelectrorealism.electrorealism.block.user.AbstractPowerUserBlock;
-import com.teamofelectrorealism.electrorealism.block.user.AbstractPowerUserBlockEntity;
+import com.teamofelectrorealism.electrorealism.block.machine.AbstractMachineBlockEntity;
+import com.teamofelectrorealism.electrorealism.block.machine.user.AbstractPowerUserBlock;
 import net.minecraft.core.BlockPos;
 import net.minecraft.network.chat.Component;
 import net.minecraft.world.InteractionHand;
@@ -27,23 +27,24 @@ import net.minecraft.world.phys.shapes.CollisionContext;
 import net.minecraft.world.phys.shapes.VoxelShape;
 import org.jetbrains.annotations.Nullable;
 
-public class ArcFurnaceBlock extends AbstractPowerUserBlock {
-    public static final DirectionProperty FACING = BlockStateProperties.HORIZONTAL_FACING;
-    public static final MapCodec<ArcFurnaceBlock> CODEC = simpleCodec(ArcFurnaceBlock::new);
-    public static final VoxelShape SHAPE = Block.box(0, 0, 0, 16, 16, 16);
+public class ElectricCrusherBlock extends AbstractPowerUserBlock {
 
-    public ArcFurnaceBlock(Properties properties) {
+    public static final DirectionProperty FACING = BlockStateProperties.HORIZONTAL_FACING;
+    public static final MapCodec<ElectricCrusherBlock> CODEC = simpleCodec(ElectricCrusherBlock::new);
+    public static final VoxelShape SHAPE = Block.box(0, 0, 0, 16, 22, 16);
+
+    public ElectricCrusherBlock(Properties properties) {
         super(properties);
     }
 
     @Override
-    protected void tick(Level level1, BlockPos pos, BlockState state1, AbstractPowerUserBlockEntity blockEntity) {
+    protected void tick(Level level1, BlockPos pos, BlockState state1, AbstractMachineBlockEntity blockEntity) {
         blockEntity.tick(level1, pos, state1);
     }
 
     @Override
     protected MapCodec<? extends BaseEntityBlock> codec() {
-        return null;
+        return CODEC;
     }
 
     @Override
@@ -73,15 +74,15 @@ public class ArcFurnaceBlock extends AbstractPowerUserBlock {
 
     @Override
     public @Nullable BlockEntity newBlockEntity(BlockPos blockPos, BlockState blockState) {
-        return new ArcFurnaceBlockEntity(blockPos, blockState);
+        return new ElectricCrusherBlockEntity(blockPos, blockState);
     }
 
     @Override
     protected void onRemove(BlockState state, Level level, BlockPos pos, BlockState newState, boolean isMoving) {
         if (state.getBlock() != newState.getBlock()) {
             BlockEntity blockEntity = level.getBlockEntity(pos);
-            if (blockEntity instanceof ArcFurnaceBlockEntity arcFurnaceBlockEntity) {
-                arcFurnaceBlockEntity.drops();
+            if (blockEntity instanceof ElectricCrusherBlockEntity electricCrusherBlockEntity) {
+                electricCrusherBlockEntity.drops();
                 level.updateNeighbourForOutputSignal(pos, this);
             }
         }
@@ -92,8 +93,8 @@ public class ArcFurnaceBlock extends AbstractPowerUserBlock {
     protected ItemInteractionResult useItemOn(ItemStack stack, BlockState state, Level level, BlockPos pos, Player player, InteractionHand hand, BlockHitResult hitResult) {
         if (!level.isClientSide()) {
             BlockEntity blockEntity = level.getBlockEntity(pos);
-            if (blockEntity instanceof ArcFurnaceBlockEntity arcFurnaceBlockEntity) {
-                player.openMenu(new SimpleMenuProvider(arcFurnaceBlockEntity, Component.literal("Arch Furnace")), pos);
+            if (blockEntity instanceof ElectricCrusherBlockEntity electricCrusherBlockEntity) {
+                player.openMenu(new SimpleMenuProvider(electricCrusherBlockEntity, Component.literal("Electric Crusher")), pos);
             } else {
                 throw new IllegalStateException("Our Container provider is missing!");
             }
@@ -107,6 +108,6 @@ public class ArcFurnaceBlock extends AbstractPowerUserBlock {
             return null;
         }
 
-        return createTickerHelper(blockEntityType, ModBlockEntityTypes.ARC_FURNACE_BE.get(), this::tick);
+        return createTickerHelper(blockEntityType, ModBlockEntityTypes.ELECTRIC_CRUSHER_BE.get(), this::tick);
     }
 }
