+++ src/main/java/com/teamofelectrorealism/electrorealism/block/machine/user/crusher/ElectricCrusherBlockEntity.java
@@ -1,13 +1,12 @@
-package com.teamofelectrorealism.electrorealism.block.user.arc_furnace;
+package com.teamofelectrorealism.electrorealism.block.machine.user.crusher;
 
 import com.teamofelectrorealism.electrorealism.api.ElectricalAPI;
-import com.teamofelectrorealism.electrorealism.block.IPowerReceiver;
 import com.teamofelectrorealism.electrorealism.block.ModBlockEntityTypes;
-import com.teamofelectrorealism.electrorealism.block.user.AbstractPowerUserBlockEntity;
+import com.teamofelectrorealism.electrorealism.block.machine.user.AbstractPowerUserBlockEntity;
 import com.teamofelectrorealism.electrorealism.recipe.ModRecipes;
-import com.teamofelectrorealism.electrorealism.recipe.arc_furnace.ArcFurnaceRecipe;
-import com.teamofelectrorealism.electrorealism.recipe.arc_furnace.ArcFurnaceRecipeInput;
-import com.teamofelectrorealism.electrorealism.screen.arc_furnace.ArcFurnaceMenu;
+import com.teamofelectrorealism.electrorealism.recipe.crusher.ElectricCrusherRecipe;
+import com.teamofelectrorealism.electrorealism.recipe.crusher.ElectricCrusherRecipeInput;
+import com.teamofelectrorealism.electrorealism.screen.crusher.ElectricCrusherMenu;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.HolderLookup;
 import net.minecraft.nbt.CompoundTag;
@@ -25,83 +24,72 @@ import net.minecraft.world.inventory.ContainerData;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.crafting.RecipeHolder;
 import net.minecraft.world.level.Level;
-import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 import net.neoforged.neoforge.items.ItemStackHandler;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.Optional;
 
-public class ArcFurnaceBlockEntity extends AbstractPowerUserBlockEntity implements MenuProvider {
-    public final ItemStackHandler itemHandler = new ItemStackHandler(2) {
+public class ElectricCrusherBlockEntity extends AbstractPowerUserBlockEntity implements MenuProvider {
+    public final ItemStackHandler itemHandler = new ItemStackHandler(3) {
         @Override
         protected void onContentsChanged(int slot) {
             if (!level.isClientSide()) {
                 setChanged();
-                level.sendBlockUpdated(getBlockPos(), getBlockState(), getBlockState(), 2);
+                level.sendBlockUpdated(getBlockPos(), getBlockState(), getBlockState(), 3);
             }
         }
     };
 
     private static final int SLOT_INPUT = 0;
-    private static final int SLOT_OUTPUT = 1;
+    private static final int SLOT_FUEL = 1;
+    private static final int SLOT_OUTPUT = 2;
 
     private static final String INVENTORY_KEY = "inventory";
-    private static final String HEAT_LEVEL_KEY = "arc_furnace.heat_level";
-    private static final String BUFFER_LEVEL_KEY = "arc_furnace.buffer_level";
-    private static final String BUFFER_TOTAL_LEVEL_KEY = "arc_furnace.buffer_total_level";
-    private static final String HEAT_TOTAL_LEVEL_KEY = "arc_furnace.heat_total_level";
-    private static final String SMELTING_PROGRESS_KEY = "arc_furnace.smelting_progress";
-    private static final String SMELTING_TOTAL_TIME_KEY = "arc_furnace.smelting_total_time";
-    private static final String INTERNAL_RESISTANCE_KEY = "arc_furnace.internal_resistance";
+    private static final String BUFFER_LEVEL_KEY = "electric_crusher.buffer_level";
+    private static final String BUFFER_TOTAL_LEVEL_KEY = "electric_crusher.buffer_total_level";
+    private static final String CRUSHING_PROGRESS_KEY = "electric_crusher.crushing_progress";
+    private static final String CRUSHING_TOTAL_TIME_KEY = "electric_crusher.crushing_total_time";
+    private static final String INTERNAL_RESISTANCE_KEY = "electric_crusher.internal_resistance";
 
-    private static final int TOTAL_SMELTING_TIME = 80;
+    private static final int TOTAL_CRUSHING_TIME = 200;
     private static final int TOTAL_BUFFER_CAPASITY = 2000; // In mAh
     private static final int INTERNAL_RESISTANCE = 10; // In ohm
 
-    private int heatLevel;
-    private int heatTotalLevel = 7;
-    private int smeltingProgress;
-    private int smeltingTotalTime = TOTAL_SMELTING_TIME;
     private int bufferLevel;
     private int bufferTotalLevel = TOTAL_BUFFER_CAPASITY;
+    private int crushingProgress;
+    private int crushingTotalTime = TOTAL_CRUSHING_TIME;
     private int internalResistance = INTERNAL_RESISTANCE;
     private final ContainerData data;
 
-    public ArcFurnaceBlockEntity(BlockPos pos, BlockState blockState) {
-        super(ModBlockEntityTypes.ARC_FURNACE_BE.get(), pos, blockState);
+    public ElectricCrusherBlockEntity(BlockPos pos, BlockState blockState) {
+        super(ModBlockEntityTypes.ELECTRIC_CRUSHER_BE.get(), pos, blockState);
         this.data = new ContainerData() {
             @Override
-            public int get(int i) {
-                return switch (i) {
-                    case 0 -> ArcFurnaceBlockEntity.this.smeltingProgress;
-                    case 1 -> ArcFurnaceBlockEntity.this.smeltingTotalTime;
-                    case 2 -> ArcFurnaceBlockEntity.this.heatLevel;
-                    case 3 -> ArcFurnaceBlockEntity.this.heatTotalLevel;
-                    case 4 -> ArcFurnaceBlockEntity.this.bufferLevel;
-                    case 5 -> ArcFurnaceBlockEntity.this.bufferTotalLevel;
-                    case 6 -> ArcFurnaceBlockEntity.this.internalResistance;
-
+            public int get(int pIndex) {
+                return switch (pIndex) {
+                    case 0 -> ElectricCrusherBlockEntity.this.crushingProgress;
+                    case 1 -> ElectricCrusherBlockEntity.this.crushingTotalTime;
+                    case 2 -> ElectricCrusherBlockEntity.this.bufferLevel;
+                    case 3 -> ElectricCrusherBlockEntity.this.bufferTotalLevel;
+                    case 4 -> ElectricCrusherBlockEntity.this.internalResistance;
                     default -> 0;
                 };
             }
-
             @Override
-            public void set(int i, int i1) {
-                switch (i) {
-                    case 0: ArcFurnaceBlockEntity.this.smeltingProgress = i;
-                    case 1: ArcFurnaceBlockEntity.this.smeltingTotalTime = i;
-                    case 2: ArcFurnaceBlockEntity.this.heatLevel = i;
-                    case 3: ArcFurnaceBlockEntity.this.heatTotalLevel = i;
-                    case 4: ArcFurnaceBlockEntity.this.bufferLevel = i;
-                    case 5: ArcFurnaceBlockEntity.this.bufferTotalLevel = i;
-                    case 6: ArcFurnaceBlockEntity.this.internalResistance = i;
-                }
+            public void set(int pIndex, int pValue) {
+                switch (pIndex) {
+                    case 0: ElectricCrusherBlockEntity.this.crushingProgress = pValue;
+                    case 1: ElectricCrusherBlockEntity.this.crushingTotalTime = pValue;
+                    case 2: ElectricCrusherBlockEntity.this.bufferLevel = pValue;
+                    case 3: ElectricCrusherBlockEntity.this.bufferTotalLevel = pValue;
+                    case 4: ElectricCrusherBlockEntity.this.internalResistance = pValue;
+                };
             }
-
             @Override
             public int getCount() {
-                return 7;
+                return 5;
             }
         };
     }
@@ -118,35 +106,31 @@ public class ArcFurnaceBlockEntity extends AbstractPowerUserBlockEntity implemen
 
     @Override
     public Component getDisplayName() {
-        return Component.translatable("block.electrorealism.arc_furnace");
+        return Component.translatable("block.electrorealism.electric_crusher");
     }
 
     @Override
     public @Nullable AbstractContainerMenu createMenu(int containerId, Inventory playerInventory, Player player) {
-        return new ArcFurnaceMenu(containerId, playerInventory, this, this.data);
+        return new ElectricCrusherMenu(containerId, playerInventory, this, this.data);
     }
 
     @Override
     public void loadAdditional(CompoundTag compoundTag, HolderLookup.Provider registries) {
         super.loadAdditional(compoundTag, registries);
         itemHandler.deserializeNBT(registries, compoundTag.getCompound(INVENTORY_KEY));
-        heatLevel = compoundTag.getInt(HEAT_LEVEL_KEY);
-        heatTotalLevel = compoundTag.getInt(HEAT_TOTAL_LEVEL_KEY);
-        smeltingProgress = compoundTag.getInt(SMELTING_PROGRESS_KEY);
-        smeltingTotalTime = compoundTag.getInt(SMELTING_TOTAL_TIME_KEY);
         bufferLevel = compoundTag.getInt(BUFFER_LEVEL_KEY);
         bufferTotalLevel = compoundTag.getInt(BUFFER_TOTAL_LEVEL_KEY);
+        crushingProgress = compoundTag.getInt(CRUSHING_PROGRESS_KEY);
+        crushingTotalTime = compoundTag.getInt(CRUSHING_TOTAL_TIME_KEY);
         internalResistance = compoundTag.getInt(INTERNAL_RESISTANCE_KEY);
     }
     @Override
     public void saveAdditional(CompoundTag compoundTag, HolderLookup.Provider registries) {
         compoundTag.put(INVENTORY_KEY, itemHandler.serializeNBT(registries));
-        compoundTag.putInt(HEAT_LEVEL_KEY, heatLevel);
-        compoundTag.putInt(HEAT_TOTAL_LEVEL_KEY, heatTotalLevel);
-        compoundTag.putInt(SMELTING_PROGRESS_KEY, smeltingProgress);
-        compoundTag.putInt(SMELTING_TOTAL_TIME_KEY, smeltingTotalTime);
         compoundTag.putInt(BUFFER_LEVEL_KEY, bufferLevel);
         compoundTag.putInt(BUFFER_TOTAL_LEVEL_KEY, bufferTotalLevel);
+        compoundTag.putInt(CRUSHING_PROGRESS_KEY, crushingProgress);
+        compoundTag.putInt(CRUSHING_TOTAL_TIME_KEY, crushingTotalTime);
         compoundTag.putInt(INTERNAL_RESISTANCE_KEY, internalResistance);
 
         super.saveAdditional(compoundTag, registries);
@@ -168,25 +152,19 @@ public class ArcFurnaceBlockEntity extends AbstractPowerUserBlockEntity implemen
     }
 
     public void tick(Level level, BlockPos blockPos, BlockState blockState) {
-        if (isPowered()) {
-            heatUp();
-        } else {
-            coolDown();
-        }
-
-        if (hasRecipe() && isOutputSlotEmptyOrReceivable()) {
-            if (isHeated() && hasBufferEnoughCharge()) {
-                increaseSmeltingProgress();
+        if(hasRecipe() && isOutputSlotEmptyOrReceivable()) {
+            if (isPowered() && hasBufferEnoughCharge()) {
+                increaseCrushingProgress();
                 decreaseBufferCharge();
-                if (hasSmeltingFinished()) {
-                    smeltItem();
+                if (hasCrushingFinished()) {
+                    crushItem();
                     resetProgress();
                 }
             } else {
-                smeltingProgress = 0;
+                crushingProgress = 0;
             }
         } else {
-            smeltingProgress = 0;
+            crushingProgress = 0;
         }
     }
 
@@ -199,7 +177,7 @@ public class ArcFurnaceBlockEntity extends AbstractPowerUserBlockEntity implemen
     }
 
     private boolean hasRecipe() {
-        Optional<RecipeHolder<ArcFurnaceRecipe>> recipe = getCurrentRecipe();
+        Optional<RecipeHolder<ElectricCrusherRecipe>> recipe = getCurrentRecipe();
 
         if (recipe.isEmpty()) {
             return false;
@@ -221,8 +199,8 @@ public class ArcFurnaceBlockEntity extends AbstractPowerUserBlockEntity implemen
         return maxCount >= currentCount + count;
     }
 
-    private Optional<RecipeHolder<ArcFurnaceRecipe>> getCurrentRecipe() {
-        return this.level.getRecipeManager().getRecipeFor(ModRecipes.ARC_FURNACE_TYPE.get(), new ArcFurnaceRecipeInput(itemHandler.getStackInSlot(SLOT_INPUT)), level);
+    private Optional<RecipeHolder<ElectricCrusherRecipe>> getCurrentRecipe() {
+        return this.level.getRecipeManager().getRecipeFor(ModRecipes.ELECTRIC_CRUSHER_TYPE.get(), new ElectricCrusherRecipeInput(itemHandler.getStackInSlot(SLOT_INPUT)), level);
     }
 
     private boolean isOutputSlotEmptyOrReceivable() {
@@ -230,12 +208,12 @@ public class ArcFurnaceBlockEntity extends AbstractPowerUserBlockEntity implemen
     }
 
     private void resetProgress() {
-        this.smeltingProgress = 0;
-        this.smeltingTotalTime = TOTAL_SMELTING_TIME;
+        this.crushingProgress = 0;
+        this.crushingTotalTime = TOTAL_CRUSHING_TIME;
     }
 
-    private void smeltItem() {
-        Optional<RecipeHolder<ArcFurnaceRecipe>> recipe = getCurrentRecipe();
+    private void crushItem() {
+        Optional<RecipeHolder<ElectricCrusherRecipe>> recipe = getCurrentRecipe();
         ItemStack output = recipe.get().value().getResultItem(null);
 
         itemHandler.extractItem(SLOT_INPUT, 1, false);
@@ -243,33 +221,16 @@ public class ArcFurnaceBlockEntity extends AbstractPowerUserBlockEntity implemen
         itemHandler.setStackInSlot(SLOT_OUTPUT, new ItemStack(output.getItem(), itemHandler.getStackInSlot(SLOT_OUTPUT).getCount() + output.getCount()));
     }
 
-    private boolean hasSmeltingFinished() {
-        return this.smeltingProgress >= this.smeltingTotalTime;
+    private boolean hasCrushingFinished() {
+        return this.crushingProgress >= this.crushingTotalTime;
     }
 
-    private void increaseSmeltingProgress() {
-        smeltingProgress++;
+    private void increaseCrushingProgress() {
+        crushingProgress++;
     }
 
-    private void increaseHeatingProgress() {
-        if (heatLevel < 7) {
-            heatLevel++;
-        }
-    }
-
-    // Timer for heating
-    private int heatUpTimer = 0;
-
-    private void heatUp() {
-        if (heatUpTimer < 160) { // 8 seconds at 20 ticks/second
-            heatUpTimer++;
-            if (heatUpTimer % 20 == 0) {
-                increaseHeatingProgress();
-            }
-        } else {
-            // Stay fully heated after 8 seconds
-            heatLevel = heatTotalLevel;
-        }
+    private boolean isPowered() {
+        return this.bufferLevel > 0;
     }
 
     @Override
@@ -294,19 +255,5 @@ public class ArcFurnaceBlockEntity extends AbstractPowerUserBlockEntity implemen
     @Override
     public void setBufferCharge(int charge) {
         this.bufferLevel = Math.max(0, Math.min(charge, bufferTotalLevel));
-    }
-
-    private void coolDown() {
-        if (heatLevel > 0) {
-            heatLevel --; // Cool down slowly
-        }
-    }
-
-    private boolean isHeated() {
-        return this.heatLevel > 6;
-    }
-
-    private boolean isPowered() {
-        return this.bufferLevel > 0;
     }
 }
