+++ src/main/java/com/teamofelectrorealism/electrorealism/screen/arc_furnace/ArcFurnaceMenu.java
@@ -1,7 +1,7 @@
 package com.teamofelectrorealism.electrorealism.screen.arc_furnace;
 
 import com.teamofelectrorealism.electrorealism.block.ModBlocks;
-import com.teamofelectrorealism.electrorealism.block.user.arc_furnace.ArcFurnaceBlockEntity;
+import com.teamofelectrorealism.electrorealism.block.machine.user.arc_furnace.ArcFurnaceBlockEntity;
 import com.teamofelectrorealism.electrorealism.screen.ModMenuTypes;
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.world.entity.player.Inventory;
